pre-submission
type: refactor
- book functionality for both cases: hp and search
- removed duplicate code
